% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingWithClasses.R
\name{generateSpecificPlots}
\alias{generateSpecificPlots}
\title{Make a specific plot}
\usage{
generateSpecificPlots(
  all_diseases,
  diseaseName,
  riskfactors,
  plot_type = c("raw", "interpolated")
)
}
\arguments{
\item{all_diseases}{Database object initialized by initialize_database}

\item{diseaseName}{the name of the disease that contains the requested plot. A string.}

\item{riskfactors}{a vector of the riskfactornames for the plot you want}

\item{plot_type}{the type of plot you want}
}
\value{
a list of plots that can later be combined to something
}
\description{
This function makes the requested plot from the database
}
