% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingWithClasses.R
\name{initialize_database}
\alias{initialize_database}
\title{Initializes the database}
\usage{
initialize_database(json_filename)
}
\arguments{
\item{json_filename}{path to the database}
}
\value{
a database object
}
\description{
This function reads the json file from the hard drive and returns a database object.
}
