% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingWithClasses.R
\name{initialize_descriptions}
\alias{initialize_descriptions}
\title{Initializes the descriptions file}
\usage{
initialize_descriptions(json_filename)
}
\arguments{
\item{json_filename}{path to the relations}
}
\value{
a descriptions object
}
\description{
This function reads the json file from the hard drive and returns a list of description objects.
}
