% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingWithClasses.R
\name{makePlot,Disease,ANY-method}
\alias{makePlot,Disease,ANY-method}
\title{Makes plots}
\usage{
\S4method{makePlot}{Disease,ANY}(x)
}
\arguments{
\item{x}{disease object which is any sub element of the database object.}
}
\value{
A list of plots.
}
\description{
This function makes a list of all the plots that are relevant to the specified disease.
}
